improve certificate request resolution
lets encrypt rate limits are pretty low for duplicate
certificate requests 5 / week. this can be problematic
with clusters undergoing significant node churn as new nodes
can initiate CSR to lets encrypt when they boot up because they
become the rendezvous node while not having a copy of the certificate
locally.

to resolve this when a node detects its the rendevzous node right
before initiating a CSR request via ACME it'll send requests for
the cached certificate to members of the extended quorum group (2 *
Quorum) this will ensure that as long as the entire extended quorum
group are not simultaneously replaced we'll be able to locate a cached
certificate instead of initiating a challenge via ACME.

We've also have the servers return not found if their cached certificate
is about to expire.

these changes combined with the earlier change to increasing the
lifetime for the cached certificate basically ensures we'll never hit
ACME with too many duplicate requests. One edge case remains as noted
about when the entire extended quorum group is replaced by new nodes
due to the rendezvous algorithms properties; but that really should
be vanishingly small.